feat: Expand internal SIMD memmem implemenation and use it everywhere
Drop the custom ad-hoc implemenations per feature and memchar::Finder
when we can use internal optimize case flip & memmem
