+++ crates/fff-core/src/dbs/mod.rs
@@ -1,4 +1,10 @@
-pub mod db_healthcheck;
-pub mod frecency;
 pub(crate) mod lmdb;
+
+pub mod db_healthcheck;
+pub use db_healthcheck::{DbHealth, DbHealthChecker};
+
+pub mod frecency;
+pub use frecency::*;
+
 pub mod query_tracker;
+pub use query_tracker::*;
