+++ crates/fff-core/src/simd_string_utils/mod.rs
@@ -1,0 +1,5 @@
+//! SIMD-accelerated string utilities: case flipping/folding and
+//! case-insensitive substring search.
+
+pub mod case;
+pub mod memmem;
